Add basic webapp demo
The user can input custom roam text, and see a pretty-printed parse
tree updated live as they type, with indication of the parse runtime.

Further down the page, they can edit `Path` and `Value` fields to
perform a live update of the internal tree structure, and see the
serialized roam string result. This update is done using built-in
Clojure functions for manipulating maps, in this case `assoc-in`.